PR#5 fixes - added public to interfaces / classes - removed ? super R in foldl / foldr - updated foldl / foldr tests
