+++ ha03/src/main/java/Collections.java
@@ -2,7 +2,8 @@ import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 
-class Collections {
+public class Collections {
+
     public static <A, R> List<R> map(Function1<? super A, R> f, Iterable<A> a) {
         List<R> res = new LinkedList<>();
         for (A elem : a) {
@@ -10,6 +11,7 @@ class Collections {
         }
         return res;
     }
+
     public static <A> List<A> filter(Predicate<? super A> p, Iterable<A> a) {
         List<A> res = new LinkedList<>();
         for (A elem : a) {
@@ -19,6 +21,7 @@ class Collections {
         }
         return res;
     }
+
     public static <A> List<A> takeWhile(Predicate<? super A> p, Iterable<A> a) {
         List<A> res = new LinkedList<>();
         for (A elem : a) {
@@ -29,31 +32,28 @@ class Collections {
         }
         return res;
     }
+
     public static <A> List<A> takeUnless(Predicate<? super A> p, Iterable<A> a) {
-        List<A> res = new LinkedList<>();
-        for (A elem : a) {
-            if (p.apply(elem)) {
-                return res;
-            }
-            res.add(elem);
-        }
-        return res;
+        return takeWhile(p.not(), a);
     }
-    public static <A, R> R foldl(Function2<? super R, ? super A, R> f, R z, Iterable<A> a) {
+
+    public static <A, R> R foldl(Function2<R, ? super A, R> f, R z, Iterable<A> a) {
         R res = z;
         for (A elem : a) {
             res = f.apply(res, elem);
         }
         return res;
     }
-    private static <A, R> R foldrRec(Function2<? super A, ? super R, R> f, R z, Iterator<A> it) {
+
+    public static <A, R> R foldr(Function2<? super A, R, R> f, R z, Iterable<A> a) {
+        return foldrRec(f, z, a.iterator());
+    }
+
+    private static <A, R> R foldrRec(Function2<? super A, R, R> f, R z, Iterator<A> it) {
         if (!it.hasNext()) {
             return z;
         }
         A elem = it.next();
         return f.apply(elem, foldrRec(f, z, it));
-    }
-    public static <A, R> R foldr(Function2<? super A, ? super R, R> f, R z, Iterable<A> a) {
-        return foldrRec(f, z, a.iterator());
     }
 }
