+++ ha03/src/main/java/Function1.java
@@ -1,0 +1,8 @@
+
+public interface Function1<A, R> {
+    R apply(A arg);
+
+    default <R1> Function1<A, R1> compose(Function1<? super R, R1> g) {
+        return arg -> g.apply(apply(arg));
+    }
+}
